Automatize grid creation and get user input
Created script to prompt the user to input grid
size. Adjusted the grid creation function to
automatize grid columns on css. It does this by
changing a root variable that was created in the
css file.
The script that changes block styles on mouseover
was also changed and put into its own function,
and now generates random colors for each block.